backend/requirements.txt
fastapi
uvicorn
scikit-learn
websockets
wsproto
pymongo
python-dotenv